Convert pumping unites to centiounces (compiles)
- This is to prevent rounding errors from using a double
--> centiounces will be used because 325 coz = 3.25 oz
- This version compiles, but is not fully functional

To do:
- Change backend over to expect an Int vs Double
- Change UI I/O to convert to centiounces < > ounces

+++ BabyPatterns/Feeding.swift
@@ -8,7 +8,7 @@ public struct Feeding {
     public var endDate: Date?
     public var lastPausedDate: Date?
     public var pausedTime: TimeInterval
-    public let supplyAmount: Double
+    public let supplyAmount: Int
 
     public let shouldPrintDebugString = true
 
@@ -25,7 +25,7 @@ public struct Feeding {
                 startDate: Date,
                 endDate: Date? = nil,
                 lastPausedDate: Date? = nil,
-                supplyAmount: Double = 0.0,
+                supplyAmount: Int = 0,
                 pausedTime: TimeInterval = 0.0,
                 serverKey: String? = nil) {
         self.type = type
@@ -48,7 +48,7 @@ public struct Feeding {
 
         let lastPausedDate = Date(timeInterval: json[K.JsonFields.LastPausedDate])
 
-        guard let supplyAmount = json[K.JsonFields.SupplyAmount] as? Double else { return nil }
+        guard let supplyAmount = json[K.JsonFields.SupplyAmount] as? Int else { return nil }
         guard let pausedTime = json[K.JsonFields.PausedTime] as? TimeInterval else { return nil }
 
         self.init(type: type,
